// src/app/services/settings/types.ts
/**
 * @author jhun.bryan.belen
 * @dateCreated 01082024
 */

export interface JOURNEY_GLOBAL_DATA {
  accountManager?: object|any
  branch?: any
  channel?: string
  location?: Array<object|any>
  occupation?: Array<object|any>
  referror?: Array<any>
  subChannel?: any
  dashBoardIndex?: number
  internet?: boolean,
  fullSyncProgress?: any,
  leadsFullSyncStatus?: any,
  naFullSyncStatus?: any,
  irpqFullSyncStatus?: any,
  siFullSyncStatus?: any,
  eappFullSyncStatus?: any,
  checklistFullSyncStatus?: any,
  referrorFullSyncStatus?: any,
  session?: any,
  requestResponse?: any,
  isLocalStorageToggle?: boolean
}